Discovery test changed
